refactor(product): 구조 수정
